Added tests for if `ServerSocket` or `ClientSocket` connection fails from Safia review
