+++ .idea/workspace.xml
@@ -4,11 +4,12 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="626fba06-9972-4b5a-b6e3-ddb2029d20d5" name="Changes" comment="Implements `readClientInputStream()` and `writeClientOutputStream()` functions and changes `EchoSocket.java` to `ClientSocket.java`">
-      <change afterPath="$PROJECT_DIR$/src/test/java/org/example/AppControllerTest.java" afterDir="false" />
+    <list default="true" id="626fba06-9972-4b5a-b6e3-ddb2029d20d5" name="Changes" comment="Add `try/catch` block to `EchoServer` from Safia's review.">
+      <change afterPath="$PROJECT_DIR$/.idea/inspectionProfiles/Project_Default.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/test/java/org/example/ClientSocketTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/EchoServer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/EchoServer.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/org/example/InputOutputTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/org/example/InputOutputTest.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/ClientSocket.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/ClientSocket.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/org/example/EchoServerTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/org/example/EchoServerTest.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -132,7 +133,21 @@
       <option name="project" value="LOCAL" />
       <updated>1669222573939</updated>
     </task>
-    <option name="localTasksCounter" value="9" />
+    <task id="LOCAL-00009" summary="Add `try/catch` block to `EchoServer` from Safia's review.">
+      <created>1669286752314</created>
+      <option name="number" value="00009" />
+      <option name="presentableId" value="LOCAL-00009" />
+      <option name="project" value="LOCAL" />
+      <updated>1669286752314</updated>
+    </task>
+    <task id="LOCAL-00010" summary="Add `try/catch` block to `EchoServer` from Safia's review.">
+      <created>1669286980660</created>
+      <option name="number" value="00010" />
+      <option name="presentableId" value="LOCAL-00010" />
+      <option name="project" value="LOCAL" />
+      <updated>1669286980660</updated>
+    </task>
+    <option name="localTasksCounter" value="11" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -158,7 +173,8 @@
     <MESSAGE value="Implements `sendSocketOutputStream(Socket socket)` function." />
     <MESSAGE value="Implements `readClientInputStream()` and `writeClientOutputStream()` functions" />
     <MESSAGE value="Implements `readClientInputStream()` and `writeClientOutputStream()` functions and changes `EchoSocket.java` to `ClientSocket.java`" />
-    <option name="LAST_COMMIT_MESSAGE" value="Implements `readClientInputStream()` and `writeClientOutputStream()` functions and changes `EchoSocket.java` to `ClientSocket.java`" />
+    <MESSAGE value="Add `try/catch` block to `EchoServer` from Safia's review." />
+    <option name="LAST_COMMIT_MESSAGE" value="Add `try/catch` block to `EchoServer` from Safia's review." />
   </component>
   <component name="XDebuggerManager">
     <breakpoint-manager>
