+++ src/main/java/org/example/ClientSocket.java
@@ -6,7 +6,11 @@ import java.net.Socket;
 
 public class ClientSocket {
 
-    public Socket socketCreator(ServerSocket socket) throws IOException {
-        return socket.accept();
+    public Socket socketCreator(ServerSocket socket) {
+        try {
+            return socket.accept();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
